contact_api_project/requirements.txt
# contact_api_project/requirements.txt
fastapi
uvicorn
python-dotenv
Jinja2
emails
httpx
pydantic[email]